+++ modin/core/execution/ray/implementations/pandas_on_ray/partitioning/partition.py
@@ -18,7 +18,11 @@ from ray.util import get_node_ip_address
 
 from modin.core.dataframe.pandas.partitioning.partition import PandasDataframePartition
 from modin.core.execution.ray.common import RayWrapper
-from modin.core.execution.ray.common.utils import ObjectIDType
+from modin.core.execution.ray.common.utils import (
+    ObjectIDType,
+    deconstruct_call_queue,
+    reconstruct_call_queue,
+)
 from modin.logging import get_logger
 from modin.pandas.indexing import compute_sliced_len
 
@@ -66,43 +70,6 @@ class PandasOnRayDataframePartition(PandasDataframePartition):
             )
         )
 
-    @staticmethod
-    def _apply_call_queue(call_queue, data):
-        """
-        Execute call queue over the given `data`.
-
-        Parameters
-        ----------
-        call_queue : list[list[func, args, kwargs], ...]
-        data : ray.ObjectRef
-
-        Returns
-        -------
-        ray.ObjectRef of pandas.DataFrame
-            The resulting pandas DataFrame.
-        ray.ObjectRef of int
-            The number of rows of the resulting pandas DataFrame.
-        ray.ObjectRef of int
-            The number of columns of the resulting pandas DataFrame.
-        ray.ObjectRef of str
-            The node IP address of the worker process.
-        """
-        (
-            num_funcs,
-            arg_lengths,
-            kw_key_lengths,
-            kw_value_lengths,
-            unfolded_queue,
-        ) = deconstruct_call_queue(call_queue)
-        return _apply_list_of_funcs.remote(
-            data,
-            num_funcs,
-            arg_lengths,
-            kw_key_lengths,
-            kw_value_lengths,
-            *unfolded_queue,
-        )
-
     def apply(self, func, *args, **kwargs):
         """
         Apply a function to the object wrapped by this partition.
@@ -134,7 +101,9 @@ class PandasOnRayDataframePartition(PandasDataframePartition):
             self._is_debug(log) and log.debug(
                 f"SUBMIT::_apply_list_of_funcs::{self._identity}"
             )
-            result, length, width, ip = self._apply_call_queue(call_queue, data)
+            result, length, width, ip = _apply_list_of_funcs.remote(
+                data, *deconstruct_call_queue(call_queue)
+            )
         else:
             # We handle `len(call_queue) == 1` in a different way because
             # this dramatically improves performance.
@@ -165,7 +134,7 @@ class PandasOnRayDataframePartition(PandasDataframePartition):
                 new_length,
                 new_width,
                 self._ip_cache,
-            ) = self._apply_call_queue(call_queue, data)
+            ) = _apply_list_of_funcs.remote(data, *deconstruct_call_queue(call_queue))
         else:
             # We handle `len(call_queue) == 1` in a different way because
             # this dramatically improves performance.
@@ -378,112 +347,6 @@ def _get_index_and_columns(df):  # pragma: no cover
         The number of columns.
     """
     return len(df.index), len(df.columns)
-
-
-def deconstruct_call_queue(call_queue):
-    """
-    Deconstruct the passed call queue into a 1D list.
-
-    This is required, so the call queue can be then passed to a Ray's kernel
-    as a variable-length argument ``kernel(*queue)`` so the Ray engine
-    automatically materialize all the futures that the queue might have contained.
-
-    Parameters
-    ----------
-    call_queue : list[list[func, args, kwargs], ...]
-
-    Returns
-    -------
-    num_funcs : int
-        The number of functions in the call queue.
-    arg_lengths : list of ints
-        The number of positional arguments for each function in the call queue.
-    kw_key_lengths : list of ints
-        The number of key-word arguments for each function in the call queue.
-    kw_value_lengths : 2D list of dict{"len": int, "was_iterable": bool}
-        Description of keyword arguments for each function. For example, `kw_value_lengths[i][j]`
-        describes the j-th keyword argument of the i-th function in the call queue.
-        The describtion contains of the lengths of the argument and whether it's a list at all
-        (for example, {"len": 1, "was_iterable": False} describes a non-list argument).
-    unfolded_queue : list
-        A 1D call queue that can be reconstructed using ``reconstruct_call_queue`` function.
-    """
-    num_funcs = len(call_queue)
-    arg_lengths = []
-    kw_key_lengths = []
-    kw_value_lengths = []
-    unfolded_queue = []
-    for call in call_queue:
-        unfolded_queue.append(call[0])
-        unfolded_queue.extend(call[1])
-        unfolded_queue.extend(call[2].keys())
-        value_lengths = []
-        for value in call[2].values():
-            was_iterable = True
-            if not isinstance(value, (list, tuple)):
-                was_iterable = False
-                value = [value]
-            unfolded_queue.extend(value)
-            value_lengths.append({"len": len(value), "was_iterable": was_iterable})
-
-        arg_lengths.append(len(call[1]))
-        kw_key_lengths.append(len(call[2]))
-        kw_value_lengths.append(value_lengths)
-
-    return num_funcs, arg_lengths, kw_key_lengths, kw_value_lengths, unfolded_queue
-
-
-def reconstruct_call_queue(
-    num_funcs, arg_lengths, kw_key_lengths, kw_value_lengths, unfolded_queue
-):
-    """
-    Reconstruct original call queue from the result of the ``deconstruct_call_queue()``.
-
-    Parameters
-    ----------
-    num_funcs : int
-        The number of functions in the call queue.
-    arg_lengths : list of ints
-        The number of positional arguments for each function in the call queue.
-    kw_key_lengths : list of ints
-        The number of key-word arguments for each function in the call queue.
-    kw_value_lengths : 2D list of dict{"len": int, "was_iterable": bool}
-        Description of keyword arguments for each function. For example, `kw_value_lengths[i][j]`
-        describes the j-th keyword argument of the i-th function in the call queue.
-        The describtion contains of the lengths of the argument and whether it's a list at all
-        (for example, {"len": 1, "was_iterable": False} describes a non-list argument).
-    unfolded_queue : list
-        A 1D call queue that is result of the ``deconstruct_call_queue()`` function.
-
-    Returns
-    -------
-    list[list[func, args, kwargs], ...]
-        Original call queue.
-    """
-    items_took = 0
-
-    def take_n_items(n):
-        nonlocal items_took
-        res = unfolded_queue[items_took : items_took + n]
-        items_took += n
-        return res
-
-    call_queue = []
-    for i in range(num_funcs):
-        func = take_n_items(1)[0]
-        args = take_n_items(arg_lengths[i])
-        kw_keys = take_n_items(kw_key_lengths[i])
-        kwargs = {}
-        value_lengths = kw_value_lengths[i]
-        for j, key in enumerate(kw_keys):
-            vals = take_n_items(value_lengths[j]["len"])
-            if value_lengths[j]["len"] == 1 and not value_lengths[j]["was_iterable"]:
-                vals = vals[0]
-            kwargs[key] = vals
-
-        call_queue.append((func, args, kwargs))
-
-    return call_queue
 
 
 @ray.remote(num_returns=4)
